% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_collection}
\alias{g3_collection}
\title{Construct a 3D geometry collection}
\usage{
g3_collection(geoms)
}
\arguments{
\item{geoms}{A list of `threesf` geometries.}
}
\value{
A `threesf` GEOMETRYCOLLECTION geometry.
}
\description{
Construct a 3D geometry collection
}
\examples{
g3_collection(list(g3_point(0, 0, 0), g3_point(1, 1, 1)))
}
