% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_from_sf}
\alias{g3_from_sf}
\title{Convert an sf geometry to threesf}
\usage{
g3_from_sf(x)
}
\arguments{
\item{x}{An `sf`, `sfc`, or `sfg` object.}
}
\value{
A single `threesf` geometry or a list of them.
}
\description{
Accepts `sf`, `sfc`, or `sfg` objects and preserves Z coordinates. Two-dimensional
input receives zero Z coordinates.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  g3_from_sf(sf::st_point(c(1, 2, 3)))
}
}
