% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_linestring}
\alias{g3_linestring}
\title{Construct a 3D linestring}
\usage{
g3_linestring(pts)
}
\arguments{
\item{pts}{A numeric matrix with two or three columns and one vertex per row.}
}
\value{
A `threesf` LINESTRING geometry.
}
\description{
Construct a 3D linestring
}
\examples{
g3_linestring(rbind(c(0, 0, 0), c(1, 1, 1)))
}
