% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_multilinestring}
\alias{g3_multilinestring}
\title{Construct 3D multilinestring geometry}
\usage{
g3_multilinestring(lines)
}
\arguments{
\item{lines}{A list of numeric coordinate matrices.}
}
\value{
A `threesf` MULTILINESTRING geometry.
}
\description{
Construct 3D multilinestring geometry
}
\examples{
g3_multilinestring(list(rbind(c(0, 0, 0), c(1, 1, 1))))
}
