% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_multipoint}
\alias{g3_multipoint}
\title{Construct a 3D multipoint}
\usage{
g3_multipoint(pts)
}
\arguments{
\item{pts}{A numeric matrix with two or three columns and one point per row.}
}
\value{
A `threesf` MULTIPOINT geometry.
}
\description{
Construct a 3D multipoint
}
\examples{
g3_multipoint(rbind(c(0, 0, 0), c(1, 1, 1)))
}
