% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_multipolygon}
\alias{g3_multipolygon}
\title{Construct a 3D multipolygon}
\usage{
g3_multipolygon(polys)
}
\arguments{
\item{polys}{A list of exterior-ring matrices or lists containing an exterior
ring followed by hole matrices.}
}
\value{
A `threesf` MULTIPOLYGON geometry.
}
\description{
Construct a 3D multipolygon
}
\examples{
g3_multipolygon(list(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
