% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{g3_plot}
\alias{g3_plot}
\title{Static or interactive 3D plotting (see man/plot.Rd).}
\usage{
g3_plot(
  g,
  interactive = FALSE,
  col = "#4477aa",
  alpha = 0.7,
  edges = TRUE,
  lwd = 1,
  point_size = 5,
  main = NULL,
  theta = 40,
  phi = 25
)
}
\description{
Static or interactive 3D plotting (see man/plot.Rd).
}
