% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_point}
\alias{g3_point}
\title{Construct a 3D point}
\usage{
g3_point(x, y, z = 0)
}
\arguments{
\item{x, y}{Numeric X and Y coordinates.}

\item{z}{Numeric Z coordinate, defaulting to zero.}
}
\value{
A `threesf` POINT geometry.
}
\description{
Construct a 3D point
}
\examples{
g3_point(1, 2, 3)
}
