% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_polygon}
\alias{g3_polygon}
\title{Construct a 3D polygon}
\usage{
g3_polygon(exterior, holes = list())
}
\arguments{
\item{exterior}{A numeric matrix defining the exterior ring.}

\item{holes}{A list of numeric matrices defining interior rings.}
}
\value{
A `threesf` POLYGON geometry.
}
\description{
Construct a 3D polygon
}
\examples{
g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0)))
}
