% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_solid}
\alias{g3_solid}
\title{Construct a 3D solid}
\usage{
g3_solid(outer, inner = list())
}
\arguments{
\item{outer}{A list of outer-shell patches.}

\item{inner}{A list of cavity shells, each represented by a list of patches.}
}
\value{
A `threesf` SOLID geometry.
}
\description{
Builds an SFCGAL-style SOLID Z from an outer shell and optional cavity shells.
}
\examples{
g3_solid(list(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
