% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_surface}
\alias{g3_surface}
\title{Construct a 3D polyhedral surface}
\usage{
g3_surface(patches)
}
\arguments{
\item{patches}{A list of ring matrices or lists containing an exterior ring
followed by hole matrices.}
}
\value{
A `threesf` POLYHEDRALSURFACE geometry.
}
\description{
Construct a 3D polyhedral surface
}
\examples{
g3_surface(list(rbind(c(0, 0, 0), c(1, 0, 0), c(0, 1, 0))))
}
