% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{g3_tin}
\alias{g3_tin}
\title{Construct a 3D triangulated irregular network}
\usage{
g3_tin(triangles)
}
\arguments{
\item{triangles}{A list of triangular ring matrices or patch lists.}
}
\value{
A `threesf` TIN geometry.
}
\description{
Construct a 3D triangulated irregular network
}
\examples{
g3_tin(list(rbind(c(0, 0, 0), c(1, 0, 0), c(0, 1, 0))))
}
