% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3darea}
\alias{st_3darea}
\title{Compute 3D area}
\usage{
st_3darea(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
The area as a numeric scalar. Holes are subtracted for planar polygons.
}
\description{
Compute 3D area
}
\examples{
st_3darea(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
