% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3darea_tessellated}
\alias{st_3darea_tessellated}
\title{Compute tessellated 3D area}
\usage{
st_3darea_tessellated(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
The tessellated surface area as a numeric scalar.
}
\description{
Compute tessellated 3D area
}
\examples{
st_3darea_tessellated(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
