% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dcentroid}
\alias{st_3dcentroid}
\title{Compute a 3D centroid}
\usage{
st_3dcentroid(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A numeric vector of length three containing X, Y, and Z coordinates.
}
\description{
Compute a 3D centroid
}
\examples{
st_3dcentroid(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
