% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dclosestpoint}
\alias{st_3dclosestpoint}
\title{Find the closest point on the first geometry}
\usage{
st_3dclosestpoint(a, b)
}
\arguments{
\item{a, b}{`threesf` geometries.}
}
\value{
A numeric vector of length three containing the closest point on `a`.
}
\description{
Find the closest point on the first geometry
}
\examples{
st_3dclosestpoint(g3_point(0, 0, 2), g3_point(0, 0, 0))
}
