% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dconvexhull}
\alias{st_3dconvexhull}
\title{Compute a 3D convex hull}
\usage{
st_3dconvexhull(g, tol = 1e-09)
}
\arguments{
\item{g}{A `threesf` geometry containing input points.}

\item{tol}{Absolute distance tolerance.}
}
\value{
A `threesf` convex hull geometry.
}
\description{
Compute a 3D convex hull
}
\examples{
st_3dconvexhull(g3_multipoint(rbind(c(0, 0, 0), c(1, 0, 0), c(0, 1, 0), c(0, 0, 1))))
}
