% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3ddfullywithin}
\alias{st_3ddfullywithin}
\title{Test whether all of one geometry is within a distance}
\usage{
st_3ddfullywithin(a, b, d)
}
\arguments{
\item{a, b}{`threesf` geometries.}

\item{d}{Non-negative distance threshold.}
}
\value{
A logical scalar indicating whether `a` is fully within `d` of `b`.
}
\description{
Test whether all of one geometry is within a distance
}
\examples{
st_3ddfullywithin(g3_point(0, 0, 1), g3_point(0, 0, 0), 1)
}
