% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3ddifference}
\alias{st_3ddifference}
\title{Subtract one 3D solid from another}
\usage{
st_3ddifference(a, b, tol = 1e-09)
}
\arguments{
\item{a, b}{`threesf` solid geometries.}

\item{tol}{Absolute distance tolerance.}
}
\value{
A `threesf` geometry containing `a` minus `b`.
}
\description{
Subtract one 3D solid from another
}
\examples{
st_3ddifference(st_extrude(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))), 0, 0, 1), st_extrude(g3_polygon(rbind(c(0.25, 0.25, 0), c(0.75, 0.25, 0), c(0.75, 0.75, 0), c(0.25, 0.75, 0))), 0, 0, 1))
}
