% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3ddistance}
\alias{st_3ddistance}
\title{Compute 3D distance}
\usage{
st_3ddistance(a, b)
}
\arguments{
\item{a, b}{`threesf` geometries.}
}
\value{
The minimum 3D distance as a numeric scalar.
}
\description{
Compute 3D distance
}
\examples{
st_3ddistance(g3_point(0.5, 0.5, 2), g3_point(0.5, 0.5, 0))
}
