% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3ddwithin}
\alias{st_3ddwithin}
\title{Test whether geometries are within a distance}
\usage{
st_3ddwithin(a, b, d)
}
\arguments{
\item{a, b}{`threesf` geometries.}

\item{d}{Non-negative distance threshold.}
}
\value{
A logical scalar.
}
\description{
Test whether geometries are within a distance
}
\examples{
st_3ddwithin(g3_point(0, 0, 1), g3_point(0, 0, 0), 1)
}
