% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dextent}
\alias{st_3dextent}
\title{Compute a 3D bounding extent}
\usage{
st_3dextent(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A numeric vector `c(xmin, ymin, zmin, xmax, ymax, zmax)`.
}
\description{
Compute a 3D bounding extent
}
\examples{
st_3dextent(g3_point(1, 2, 3))
}
