% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dintersection}
\alias{st_3dintersection}
\title{Intersect two 3D solids}
\usage{
st_3dintersection(a, b, tol = 1e-09)
}
\arguments{
\item{a, b}{`threesf` solid geometries.}

\item{tol}{Absolute distance tolerance.}
}
\value{
A `threesf` geometry containing the intersection.
}
\description{
Intersect two 3D solids
}
\examples{
st_3dintersection(st_extrude(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))), 0, 0, 1), st_extrude(g3_polygon(rbind(c(0.5, 0.5, 0), c(1.5, 0.5, 0), c(1.5, 1.5, 0), c(0.5, 1.5, 0))), 0, 0, 1))
}
