% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dintersects}
\alias{st_3dintersects}
\title{Test whether geometries intersect}
\usage{
st_3dintersects(a, b, eps = 1e-09)
}
\arguments{
\item{a, b}{`threesf` geometries.}

\item{eps}{Numeric intersection tolerance.}
}
\value{
A logical scalar.
}
\description{
Test whether geometries intersect
}
\examples{
st_3dintersects(g3_point(0, 0, 0), g3_point(0, 0, 0))
}
