% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dlength}
\alias{st_3dlength}
\title{Compute 3D length}
\usage{
st_3dlength(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
The 3D length as a numeric scalar.
}
\description{
Compute 3D length
}
\examples{
st_3dlength(g3_linestring(rbind(c(0, 0, 0), c(1, 2, 2))))
}
