% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dlongestline}
\alias{st_3dlongestline}
\title{Find the longest connecting line}
\usage{
st_3dlongestline(a, b)
}
\arguments{
\item{a, b}{`threesf` geometries.}
}
\value{
A 2 by 3 numeric matrix containing the farthest points on `a` and `b`.
}
\description{
Find the longest connecting line
}
\examples{
st_3dlongestline(g3_point(0, 0, 2), g3_point(0, 0, 0))
}
