% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dmaxdistance}
\alias{st_3dmaxdistance}
\title{Compute maximum 3D distance}
\usage{
st_3dmaxdistance(a, b)
}
\arguments{
\item{a, b}{`threesf` geometries.}
}
\value{
The maximum 3D distance as a numeric scalar.
}
\description{
Compute maximum 3D distance
}
\examples{
st_3dmaxdistance(g3_point(0, 0, 0), g3_point(1, 1, 1))
}
