% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dperimeter}
\alias{st_3dperimeter}
\title{Compute 3D perimeter}
\usage{
st_3dperimeter(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
The 3D perimeter as a numeric scalar.
}
\description{
Compute 3D perimeter
}
\examples{
st_3dperimeter(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
