% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_3dshortestline}
\alias{st_3dshortestline}
\title{Find the shortest connecting line}
\usage{
st_3dshortestline(a, b)
}
\arguments{
\item{a, b}{`threesf` geometries.}
}
\value{
A 2 by 3 numeric matrix: the closest point on `a` in row one and
  the closest point on `b` in row two.
}
\description{
Find the shortest connecting line
}
\examples{
st_3dshortestline(g3_point(0, 0, 2), g3_point(0, 0, 0))
}
