% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_asbinary}
\alias{st_asbinary}
\title{Convert a geometry to WKB}
\usage{
st_asbinary(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A raw vector containing WKB without an SRID.
}
\description{
Convert a geometry to WKB
}
\examples{
st_asbinary(st_geomfromtext("POINT Z (1 2 3)"))
}
