% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_asewkb}
\alias{st_asewkb}
\title{Convert a geometry to EWKB}
\usage{
st_asewkb(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A raw vector containing EWKB, including an SRID when set.
}
\description{
Convert a geometry to EWKB
}
\examples{
st_asewkb(st_setsrid(st_geomfromtext("POINT Z (1 2 3)"), 4326))
}
