% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_asewkt}
\alias{st_asewkt}
\title{Convert a geometry to EWKT}
\usage{
st_asewkt(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A character string containing EWKT, including an SRID when set.
}
\description{
Convert a geometry to EWKT
}
\examples{
st_asewkt(st_setsrid(st_geomfromtext("POINT Z (1 2 3)"), 4326))
}
