% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_ashexewkb}
\alias{st_ashexewkb}
\title{Convert a geometry to hexadecimal EWKB}
\usage{
st_ashexewkb(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A hexadecimal character representation of EWKB.
}
\description{
Convert a geometry to hexadecimal EWKB
}
\examples{
st_ashexewkb(st_geomfromtext("POINT Z (1 2 3)"))
}
