% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_astext}
\alias{st_astext}
\title{Convert a geometry to WKT}
\usage{
st_astext(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A character string containing WKT without an SRID prefix.
}
\description{
Convert a geometry to WKT
}
\examples{
st_astext(st_geomfromtext("POINT Z (1 2 3)"))
}
