% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_extrude}
\alias{st_extrude}
\title{Extrude a polygon into a solid}
\usage{
st_extrude(g, dx, dy, dz)
}
\arguments{
\item{g}{A `threesf` polygon geometry.}

\item{dx, dy, dz}{Numeric extrusion distances along the three axes.}
}
\value{
A `threesf` SOLID geometry.
}
\description{
Sweeps the first polygon of `g` along the vector `(dx, dy, dz)`.
}
\examples{
st_extrude(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))), 0, 0, 1)
}
