% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_force_outward}
\alias{st_force_outward}
\title{Orient a solid outward}
\usage{
st_force_outward(g)
}
\arguments{
\item{g}{A `threesf` solid or closed surface geometry.}
}
\value{
A `threesf` geometry with outward-oriented patches.
}
\description{
Orient a solid outward
}
\examples{
st_force_outward(st_extrude(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))), 0, 0, 1))
}
