% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_forcepolygonccw}
\alias{st_forcepolygonccw}
\title{Force polygon rings counterclockwise}
\usage{
st_forcepolygonccw(g)
}
\arguments{
\item{g}{A `threesf` polygon geometry.}
}
\value{
A `threesf` geometry with counterclockwise exterior orientation.
}
\description{
Force polygon rings counterclockwise
}
\examples{
st_forcepolygonccw(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
