% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_forcepolygoncw}
\alias{st_forcepolygoncw}
\title{Force polygon rings clockwise}
\usage{
st_forcepolygoncw(g)
}
\arguments{
\item{g}{A `threesf` polygon geometry.}
}
\value{
A `threesf` geometry with clockwise exterior orientation.
}
\description{
Force polygon rings clockwise
}
\examples{
st_forcepolygoncw(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
