% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_geometrytype}
\alias{st_geometrytype}
\title{Get the geometry type}
\usage{
st_geometrytype(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A character scalar such as `"POINT"` or `"SOLID"`.
}
\description{
Get the geometry type
}
\examples{
st_geometrytype(st_geomfromtext("POINT Z (1 2 3)"))
}
