% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_geomfromtext}
\alias{st_geomfromtext}
\title{Read a geometry from WKT or EWKT}
\usage{
st_geomfromtext(wkt)
}
\arguments{
\item{wkt}{A WKT or EWKT string.}
}
\value{
A `threesf` geometry external pointer.
}
\description{
Parses a WKT or EWKT string into a native `threesf` geometry. Two-dimensional
input is promoted to Z coordinates with zero elevation.
}
\examples{
st_geomfromtext("POINT Z (1 2 3)")
}
