% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_geomfromwkb}
\alias{st_geomfromwkb}
\title{Read a geometry from WKB or EWKB}
\usage{
st_geomfromwkb(wkb)
}
\arguments{
\item{wkb}{A raw WKB/EWKB vector or hexadecimal WKB/EWKB string.}
}
\value{
A `threesf` geometry external pointer.
}
\description{
Accepts either a raw vector or a hexadecimal string.
}
\examples{
st_geomfromwkb(st_asbinary(st_geomfromtext("POINT Z (1 2 3)")))
}
