% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_isclosed}
\alias{st_isclosed}
\title{Test whether a surface is closed}
\usage{
st_isclosed(g)
}
\arguments{
\item{g}{A `threesf` surface geometry.}
}
\value{
A logical scalar. A closed surface has every edge shared by two patches.
}
\description{
Test whether a surface is closed
}
\examples{
st_isclosed(g3_linestring(rbind(c(0, 0, 0), c(1, 1, 1))))
}
