% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_isempty}
\alias{st_isempty}
\title{Test whether a geometry is empty}
\usage{
st_isempty(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A logical scalar.
}
\description{
Test whether a geometry is empty
}
\examples{
st_isempty(st_geomfromtext("POINT Z EMPTY"))
}
