% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_isplanar}
\alias{st_isplanar}
\title{Test whether a geometry is planar}
\usage{
st_isplanar(g, tol = 1e-09)
}
\arguments{
\item{g}{A `threesf` geometry.}

\item{tol}{Maximum vertex distance from the best-fit plane.}
}
\value{
A logical scalar.
}
\description{
Test whether a geometry is planar
}
\examples{
st_isplanar(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
