% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_issolid}
\alias{st_issolid}
\title{Test whether a geometry is a solid}
\usage{
st_issolid(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
A logical scalar indicating whether `g` is a consistently oriented solid.
}
\description{
Test whether a geometry is a solid
}
\examples{
st_issolid(st_extrude(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))), 0, 0, 1))
}
