% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_makesolid}
\alias{st_makesolid}
\title{Make a solid from a closed surface}
\usage{
st_makesolid(g)
}
\arguments{
\item{g}{A closed `threesf` surface geometry.}
}
\value{
A `threesf` SOLID geometry.
}
\description{
Make a solid from a closed surface
}
\examples{
st_makesolid(st_extrude(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))), 0, 0, 1))
}
