% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_npoints}
\alias{st_npoints}
\title{Count geometry vertices}
\usage{
st_npoints(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
An integer vertex count.
}
\description{
Count geometry vertices
}
\examples{
st_npoints(st_geomfromtext("LINESTRING Z (0 0 0, 1 1 1)"))
}
