% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_orientation}
\alias{st_orientation}
\title{Get polygon orientation}
\usage{
st_orientation(g)
}
\arguments{
\item{g}{A `threesf` polygon geometry.}
}
\value{
`1` for counterclockwise orientation, `-1` for clockwise orientation,
  or `0` when orientation is undefined.
}
\description{
Get polygon orientation
}
\examples{
st_orientation(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
