% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_selfintersects}
\alias{st_selfintersects}
\title{Test for self-intersections}
\usage{
st_selfintersects(g, tol = 1e-09)
}
\arguments{
\item{g}{A `threesf` geometry.}

\item{tol}{Absolute distance tolerance.}
}
\value{
A logical scalar.
}
\description{
Test for self-intersections
}
\examples{
st_selfintersects(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))))
}
