% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_setsrid}
\alias{st_setsrid}
\title{Set a geometry SRID}
\usage{
st_setsrid(g, srid)
}
\arguments{
\item{g}{A `threesf` geometry.}

\item{srid}{An integer spatial reference identifier.}
}
\value{
A `threesf` geometry with the requested SRID.
}
\description{
Set a geometry SRID
}
\examples{
st_setsrid(st_geomfromtext("POINT Z (1 2 3)"), 4326)
}
