% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_signed_volume}
\alias{st_signed_volume}
\title{Compute signed solid volume}
\usage{
st_signed_volume(g)
}
\arguments{
\item{g}{A closed `threesf` surface or solid geometry.}
}
\value{
The signed volume as a numeric scalar; orientation determines its sign.
}
\description{
Compute signed solid volume
}
\examples{
st_signed_volume(st_extrude(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))), 0, 0, 1))
}
