% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_srid}
\alias{st_srid}
\title{Get a geometry SRID}
\usage{
st_srid(g)
}
\arguments{
\item{g}{A `threesf` geometry.}
}
\value{
An integer SRID, or zero when no SRID is set.
}
\description{
Get a geometry SRID
}
\examples{
st_srid(st_geomfromtext("SRID=4326;POINT Z (1 2 3)"))
}
