% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_tesselate}
\alias{st_tesselate}
\title{Tessellate a geometry}
\usage{
st_tesselate(g)
}
\arguments{
\item{g}{A `threesf` surface or solid geometry.}
}
\value{
A `threesf` TIN geometry.
}
\description{
Converts surface content into a TIN Z geometry.
}
\examples{
st_tesselate(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(0, 1, 0))))
}
