% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesf.R
\name{st_volume}
\alias{st_volume}
\title{Compute solid volume}
\usage{
st_volume(g)
}
\arguments{
\item{g}{A closed `threesf` surface or solid geometry.}
}
\value{
The volume as a numeric scalar.
}
\description{
Compute solid volume
}
\examples{
st_volume(st_extrude(g3_polygon(rbind(c(0, 0, 0), c(1, 0, 0), c(1, 1, 0), c(0, 1, 0))), 0, 0, 1))
}
